+++ unit/api/application_form_resource.py
@@ -1,5 +1,5 @@
 from unit.api.base_resource import BaseResource
-from unit.models.applicationForm import *
+from unit.models.application_form import *
 from unit.models.codecs import DtoDecoder
